stealing/dropzone_requirements.txt

webull
psycopg2-binary==2.8.6
pyttsx3
pretty_errors
cufflinks
pandas_ta
pandas_datareader
fastquant
historic_crypto
tweepy==3.8.0

